separate laod methods into diferent file
